Move config file reading to another function for testing
