+++ main.go
@@ -3,8 +3,6 @@ package main
 import (
 	"flag"
 	"fmt"
-	"gopkg.in/yaml.v3"
-	"io"
 	"log"
 	"os"
 	"path"
@@ -39,53 +37,8 @@ func main() {
 		Errors:    make([]*Error, 0),
 		RWMutex:   new(sync.RWMutex),
 	}
-	// if config_file is default, see if there's an `.ls-lint.yaml` file instead
-	if *config_file == ".ls-lint.yml" {
-		files, err := os.ReadDir(cwd)
-		if err != nil {
-			log.Fatal(err)
-		}
 	
-		for _, file := range files {
-			match, match_err := path.Match(".ls-lint.y*ml", file.Name())
-			if match_err != nil {
-				log.Fatal(match_err)
-			}
-			if match {
-				*config_file = file.Name()
-				break
-			}
-		}
-	}
-	// open config file
-	file, err := os.Open(*config_file)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// close file
-	defer func() {
-		err = file.Close()
-
-		if err != nil {
-			log.Fatal(err)
-		}
-	}()
-
-	// read file
-	configBytes, err := io.ReadAll(file)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// to yaml
-	err = yaml.Unmarshal(configBytes, &config)
-
-	if err != nil {
-		log.Fatal(err)
-	}
+	read_config_file(cwd, *config_file, config)
 
 	// runner
 	if err := linter.Run(filesystem, config, *debug, false); err != nil {
